Added user registration
Users can now be successfully created.  However, their password is not yet stored as a hash, as we currently have no idea how to use the encoder.

Also made some User columns non nullable, and forced nicknames to be unique. However, the exception raised when a user is saved with an existing nickname is not handled.

Added a simple script to register.mustache to check that the passwords are the same.

Co-Authored-By: Lucía Domínguez Rodrigo <117301166+LuciaDominguezRodrigo@users.noreply.github.com>