validation of fields has been created when creating an event
For this, a minimum of 1 has been established for capacity and 0 for price. It is verified that the start date of the event is after the current day and that the end date is after the start date. Furthermore, it begins that the coordinates follow a number comma number pattern.