Error in delete account fixed
If the profile was edited and an attempt was made to delete the account, an error occurred because the credentials fetched in the delete account function were those of the initially logged-in user. Now, when editing the data, the user is forced to redirect to the login page to access with the new credentials. It is worth noting that a warning pops up, notifying the user that they will need to perform this action when modifying their data.