Fixed an error caused by default reviews
Changes to the DatabaseInitializer done in commit titled "Event Ticket Visualizer and PDF Generation" were causing issues. This commit undoes those changes. Apparently, there was some sort of circular dependency or similar problem.

We found that not instantiating these reviews temporarily solves the problem. So, in order to be able to properly test other functionalities, this change will be reverted until the underlying issue is solved.

Co-Authored-By: Lucía Domínguez Rodrigo <117301166+LuciaDominguezRodrigo@users.noreply.github.com>
Co-Authored-By: Tarek Elshami <117302441+TarekElshami@users.noreply.github.com>