Now the edit and delete event buttons work
The same mustacha has been used to create events and edit them, ensuring full functionality in both. The photo of the event cannot be initialized and therefore whenever the user enters to edit an event they will have to enter the photo. When you delete an event, it disappears from the database unless it has a review assigned.