Improving the more info screen and solving the error when deleting a user
The new section of the screen for more information about an event has been built

Now if the event is finished the new section appears, otherwise the actions section appears.

If the user is the creator, you can enter the number of users who attended your event and see the graph

If you are not the creator, you will be able to rate that event.

In both cases (being a creator or not) you will be able to see the average rating

In addition, that page has been decorated a little

Fix a bug that deleted events the user was registered for if they deleted their account