Solved an issue where clicking buttons in profile would do nothing
The lines causing the issue were added in a previous commit. As I didn't do that commit or participate on it, I do not know their function.

The lines have been commented, to make it easier to figure out why they were causing issues.

IMPORTANT: this should not be left alone. Whatever those lines did still needs to be done, so it should be investigated and done properly.