Now you can record the number of people who have attended your event
If you are the event creator or admin, you can record the number of people who have attended the event and see the graph comparing the people who have attended with those who have not.

The security of the website has also been improved, preventing you from being able to add attendance for an event that is not yours.

If you try to view the graph without having indicated attendance, all values ​​will be at -1