An EventFinishedDTO has been added
EventFinishedDTO inherits from EventDTO so one type or another can be returned depending on the date without changing the return.

I have added to the EventFinishedDTO the average rating, the number of reviews and also the number of people who have attended.

I have also added the duration of the event to the EventDTO

I have changed the path to create an event from /api/events/ to /api/events without the second /

Finally I have modified the web part to show the number of reviews in addition to the average rating