The image has been separated from the event when creating it
Before: You could create the event with a photo in the same REST request.

Now: You need to make a POST with the event data and then a PUT to put the photo.

In addition, when using JwtRequestFilter it is no longer necessary to control the 401 unauthorized, so all the end points that used it have been changed with their documentation.